+++ src/components/marginals/index.js
@@ -1,3 +1,4 @@
 
 export { default as Footer } from "./Footer/Footer";
-export {default as Navbar } from "./Navbar/Navigation"
+export { default as Navbar } from "./Navbar/Navigation";
+export { default as SearchBar } from "./Search/Search";
